fix: freeCam while Sodium is installed by isolating the offending Mixin using Conditional Mixins .
