+++ src/main/java/fi/dy/masa/tweakeroo/mixin/freecam/MixinItemInHandRenderer_freeCam.java
@@ -14,7 +14,7 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 
-@Mixin(value = ItemInHandRenderer.class, priority = 1005)
+@Mixin(value = ItemInHandRenderer.class, priority = 850)
 public abstract class MixinItemInHandRenderer_freeCam
 {
     @Inject(method = "renderItem(Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/item/ItemDisplayContext;Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;I)V",
