+++ src/main/java/fi/dy/masa/tweakeroo/mixin/freecam/MixinMultiPlayerGameMode_freeCam.java
@@ -9,7 +9,7 @@ import fi.dy.masa.tweakeroo.config.Configs;
 import fi.dy.masa.tweakeroo.config.FeatureToggle;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 
-@Mixin(value = MultiPlayerGameMode.class, priority = 1005)
+@Mixin(value = MultiPlayerGameMode.class, priority = 850)
 public class MixinMultiPlayerGameMode_freeCam
 {
 	@Inject(method = "hasExperience", at = @At("RETURN"), cancellable = true)
