+++ src/main/resources/mixins.tweakeroo.json
@@ -17,7 +17,7 @@
         "block.MixinSculkSensor",
         "block.MixinSignBlockEntity",
         "block.MixinSlimeBlock",
-        "block.MixinStructureBlockEntity",
+        "block.carpet_breaks.MixinStructureBlockEntity",
         "easyplace.MixinBlockItem_easyPlace",
         "easyplace.MixinServerGamePacketListenerImpl_easyPlace",
         "entity.MixinAttackRange",
@@ -41,7 +41,7 @@
         "item.MixinItem",
         "item.MixinItemStack",
         "item.MixinSpyglassItem",
-        "network.MixinServerboundSetStructureBlockPacket",
+        "network.carpet_breaks.MixinServerboundSetStructureBlockPacket",
         "world.IMixinLightEngine",
         "world.MixinChunkMap",
         "world.MixinLevel",
@@ -56,9 +56,9 @@
         "fog.MixinAtmosphericFogEnvironment",
         "fog.MixinDarknessFogEnvironment",
         "fog.MixinFogRenderer",
-        "fog.MixinGameRenderer_disableFog",
         "fog.MixinLavaFogEnvironment",
         "fog.MixinWaterFogEnvironment",
+        "fog.sodium_breaks.MixinGameRenderer_disableFog_sodiumBreaks",
         "freecam.MixinCamera_freeCam",
         "freecam.MixinGameRenderer_freeCam",
         "freecam.MixinHud_freeCam",
@@ -68,6 +68,7 @@
         "freecam.MixinLocalPlayer_freeCam",
         "freecam.MixinMinecraft_freeCam",
         "freecam.MixinMultiPlayerGameMode_freeCam",
+        "freecam.sodium_breaks.MixinLevelRenderer_freeCam_sodiumBreaks",
         "hud.MixinBossHealthOverlay",
         "hud.MixinChatComponent",
         "hud.MixinEffectsInInventory",
